% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodchlorinationsurvey.R
\docType{data}
\name{floodchlorinationsurvey}
\alias{floodchlorinationsurvey}
\title{USAID Flood Response Rehabilitation and Chlorination Survey}
\format{
A tibble with 328 rows and 118 variables
\describe{
\item{submitted_on}{Date when the data was submitted}
\item{water_point_name}{Name of the water point}
\item{latitude}{Latitude coordinate of the water point}
\item{longitude}{Longitude coordinate of the water point}
\item{chlorination_rehab_date}{Date chlorination or rehabilitation was conducted}
\item{community_mobilised}{Indicates whether the community was mobilized for support}
\item{functional_status_before}{Functional status of the water point before intervention}
\item{rating_bolts_nuts}{Condition rating of bolts and nuts}
\item{rating_compression_cone}{Condition rating of the compression cone}
\item{rating_cylinder}{Condition rating of the cylinder}
\item{rating_flapper}{Condition rating of the flapper}
\item{rating_footvalve_bobbin}{Condition rating of the footvalve bobbin}
\item{rating_footvalve_body}{Condition rating of the footvalve body}
\item{rating_footvalve_oring}{Condition rating of the footvalve O ring}
\item{rating_fulcrum_pin}{Condition rating of the fulcrum pin}
\item{rating_fulcrum_bearings}{Condition rating of the fulcrum pin bush bearings}
\item{rating_handle}{Condition rating of the pump handle}
\item{rating_hanger_pin}{Condition rating of the hanger pin}
\item{rating_hanger_bearings}{Condition rating of the hanger pin bush bearings}
\item{rating_head_bolt}{Condition rating of the head bolt}
\item{rating_plunger_bobbin}{Condition rating of the plunger bobbin}
\item{rating_plunger_body}{Condition rating of the plunger body}
\item{rating_plunger_useal}{Condition rating of the plunger U seal}
\item{rating_pump_head_case}{Condition rating of the pump head case}
\item{rating_pump_head_cover}{Condition rating of the pump head cover}
\item{rating_rising_main}{Condition rating of the rising main pipe}
\item{rating_rising_centralisers}{Condition rating of rising main centralisers}
\item{rating_rods}{Condition rating of connecting rods}
\item{rating_rod_centralisers}{Condition rating of rod centralisers}
\item{rating_rodhanger}{Condition rating of the rod hanger}
\item{rating_rope}{Condition rating of the rope}
\item{rating_socket}{Condition rating of the socket}
\item{rating_steel_cone}{Condition rating of the steel cone}
\item{rating_top_sleeve}{Condition rating of the top sleeve}
\item{rod_count}{Number of rods in the borehole}
\item{broken_parts_fished_out}{Indicates whether broken parts were retrieved from the borehole}
\item{broken_parts_comments}{Comments on the broken parts retrieved}
\item{silt_observed}{Presence of silt on parts (Yes/No)}
\item{corrosion_observed}{Presence of corrosion on parts (Yes/No)}
\item{slime_observed}{Presence of slime buildup on parts (Yes/No)}
\item{rising_main_length_m}{Length of the rising main pipe in meters}
\item{rope_attached}{Indicates whether a rope was attached to the rising main}
\item{photos_of_removed_parts}{URLs or filenames of photos showing removed parts}
\item{pedestal_condition}{General condition of the pedestal}
\item{pedestal_firm}{Is the pedestal held firmly in place?}
\item{photos_of_civil_work}{URLs or filenames of photos showing surrounding infrastructure}
\item{latrines_within_30m}{Presence of latrines within 30m of the water point}
\item{latrines_within_100m}{Presence of latrines within 100m of the water point}
\item{cemetery_within_30m}{Presence of cemetery within 30m}
\item{waste_dumps_within_30m}{Presence of waste dumps within 30m}
\item{difficult_access_by_car}{Whether the site is difficult to reach by car}
\item{chlorination_done}{Indicates whether chlorination was conducted}
\item{parts_chlorinated}{Indicates whether the parts were chlorinated}
\item{parts_chlorinated_unknown}{TRUE if unsure whether parts were chlorinated}
\item{borehole_depth_m}{Total depth of the borehole in meters}
\item{water_depth_m}{Depth of water in the borehole in meters}
\item{chlorine_water_buckets}{Number of 20L buckets of chlorine-treated water used}
\item{hth_dissolved_properly}{Whether 1g of HTH was dissolved for every 20L of water}
\item{chlorinated_parts_installed}{Were chlorinated parts reinstalled after treatment?}
\item{replaced_rising_main_length_m}{Length of rising main pipe replaced (m)}
\item{replaced_rod_count}{Number of rods replaced}
\item{replaced_bolts_nuts_tick}{TRUE if bolts and nuts were replaced}
\item{replaced_bolts_nuts_count}{Number of bolts and nuts replaced}
\item{replaced_compression_cone_tick}{TRUE if compression cone was replaced}
\item{replaced_compression_cone_count}{Number of compression cones replaced}
\item{replaced_cylinder_tick}{TRUE if cylinder was replaced}
\item{replaced_cylinder_count}{Number of cylinders replaced}
\item{replaced_flapper_tick}{TRUE if flapper was replaced}
\item{replaced_flapper_count}{Number of flappers replaced}
\item{replaced_footvalve_bobbin_tick}{TRUE if footvalve bobbin was replaced}
\item{replaced_footvalve_bobbin_count}{Number of footvalve bobbins replaced}
\item{replaced_footvalve_body_tick}{TRUE if footvalve body was replaced}
\item{replaced_footvalve_body_count}{Number of footvalve bodies replaced}
\item{replaced_footvalve_oring_tick}{TRUE if footvalve O-ring was replaced}
\item{replaced_footvalve_oring_count}{Number of O-rings replaced}
\item{replaced_fulcrum_pin_tick}{TRUE if fulcrum pin was replaced}
\item{replaced_fulcrum_pin_count}{Number of fulcrum pins replaced}
\item{replaced_fulcrum_bearings_tick}{TRUE if fulcrum bearings were replaced}
\item{replaced_fulcrum_bearings_count}{Number of fulcrum bearings replaced}
\item{replaced_handle_tick}{TRUE if handle was replaced}
\item{replaced_handle_count}{Number of handles replaced}
\item{replaced_hanger_pin_tick}{TRUE if hanger pin was replaced}
\item{replaced_hanger_pin_count}{Number of hanger pins replaced}
\item{replaced_hanger_bearings_tick}{TRUE if hanger bearings were replaced}
\item{replaced_hanger_bearings_count}{Number of hanger bearings replaced}
\item{replaced_head_bolt_tick}{TRUE if head bolt was replaced}
\item{replaced_head_bolt_count}{Number of head bolts replaced}
\item{replaced_plunger_bobbin_tick}{TRUE if plunger bobbin was replaced}
\item{replaced_plunger_bobbin_count}{Number of plunger bobbins replaced}
\item{replaced_plunger_body_tick}{TRUE if plunger body was replaced}
\item{replaced_plunger_body_count}{Number of plunger bodies replaced}
\item{replaced_plunger_useal_tick}{TRUE if plunger U-seal was replaced}
\item{replaced_plunger_useal_count}{Number of U-seals replaced}
\item{replaced_pump_head_case_tick}{TRUE if pump head case was replaced}
\item{replaced_pump_head_case_count}{Number of pump head cases replaced}
\item{replaced_pump_head_cover_tick}{TRUE if pump head cover was replaced}
\item{replaced_pump_head_cover_count}{Number of pump head covers replaced}
\item{replaced_rising_main_tick}{TRUE if rising main pipe was replaced}
\item{replaced_rising_main_count}{Number of rising main pipes replaced}
\item{replaced_rising_centralisers_tick}{TRUE if rising main centralisers were replaced}
\item{replaced_rising_centralisers_count}{Number of centralisers replaced}
\item{replaced_rods_tick}{TRUE if rods were replaced}
\item{replaced_rods_count}{Number of rods replaced}
\item{replaced_rod_centralisers_tick}{TRUE if rod centralisers were replaced}
\item{replaced_rod_centralisers_count}{Number of rod centralisers replaced}
\item{replaced_rodhanger_tick}{TRUE if rodhanger was replaced}
\item{replaced_rodhanger_count}{Number of rodhangers replaced}
\item{replaced_rope_tick}{TRUE if rope was replaced}
\item{replaced_rope_count}{Number of ropes replaced}
\item{replaced_socket_tick}{TRUE if socket was replaced}
\item{replaced_socket_count}{Number of sockets replaced}
\item{replaced_steel_cone_tick}{TRUE if steel cone was replaced}
\item{replaced_steel_cone_count}{Number of steel cones replaced}
\item{replaced_top_sleeve_tick}{TRUE if top sleeve was replaced}
\item{replaced_top_sleeve_count}{Number of top sleeves replaced}
\item{functionality_after}{Functional status of the water point after intervention}
\item{time_to_fill_20l_bucket}{Time (in seconds) it took to fill a 20L bucket}
\item{strokes_to_yield_water}{Number of strokes required to yield water from the pump}
}
}
\usage{
floodchlorinationsurvey
}
\description{
This dataset is from the USAID Flood Response Rehabilitation and Chlorination Survey, conducted between 2019 - 2020, to assess and rehabilitate flood-affected
boreholes in Malawi. Using the mWater, data collectors captured detailed information on the condition of handpumps, their components, water safety risks, and
the impact of chlorination and repairs. The survey was part of an emergency response initiative to restore access to safe drinking water following flood events.
}
\keyword{datasets}
